Refactor Navigation Module
